+++ packages/core/src/types/PdfJs.ts
@@ -52,6 +52,7 @@ export declare namespace PdfJs {
         getOutline(): Promise<Outline[]>;
         getPage(pageIndex: number): Promise<Page>;
         getPageIndex(ref: OutlineRef): Promise<number>;
+        getPageLabels(): Promise<string[] | null>;
         getPermissions(): Promise<number[] | null>;
     }
     interface GetDocumentParams {