+++ packages/thumbnail/src/ThumbnailList.tsx
@@ -13,15 +13,17 @@ import type { PdfJs } from '@react-pdf-viewer/core';
 import { LabelsLoader } from './LabelsLoader';
 import { scrollToBeVisible } from './scrollToBeVisible';
 import { ThumbnailContainer } from './ThumbnailContainer';
+import type { RenderCurrentPageLabel } from './types/RenderCurrentPageLabelProps';
 
 export const ThumbnailList: React.FC<{
     currentPage: number;
     doc: PdfJs.PdfDocument;
     pageHeight: number;
     pageWidth: number;
+    renderCurrentPageLabel?: RenderCurrentPageLabel;
     rotation: number;
     onJumpToPage(pageIndex: number): void;
-}> = ({ currentPage, doc, pageHeight, pageWidth, rotation, onJumpToPage }) => {
+}> = ({ currentPage, doc, pageHeight, pageWidth, renderCurrentPageLabel, rotation, onJumpToPage }) => {
     const { numPages } = doc;
     const containerRef = React.useRef<HTMLDivElement | null>(null);
     const thumbnailsRef = React.useRef<HTMLElement[]>([]);
@@ -133,32 +135,37 @@ export const ThumbnailList: React.FC<{
                     >
                         {Array(numPages)
                             .fill(0)
-                            .map((_, index) => (
-                                <div key={`thumbnail-${index}`}>
-                                    <div
-                                        className={classNames({
-                                            'rpv-thumbnail__item': true,
-                                            'rpv-thumbnail__item--selected': currentPage === index,
-                                        })}
-                                        role="button"
-                                        tabIndex={currentPage === index ? 0 : -1}
-                                        onClick={() => onJumpToPage(index)}
-                                    >
-                                        <ThumbnailContainer
-                                            doc={doc}
-                                            isActive={currentPage === index}
-                                            pageHeight={pageHeight}
-                                            pageIndex={index}
-                                            pageWidth={pageWidth}
-                                            rotation={rotation}
-                                            onActive={scrollToThumbnail}
-                                        />
+                            .map((_, pageIndex) => {
+                                const pageLabel = numLabels === numPages ? labels[pageIndex] : `${pageIndex + 1}`;
+                                const label = renderCurrentPageLabel
+                                    ? renderCurrentPageLabel({ currentPage, pageIndex, numPages, pageLabel })
+                                    : pageLabel;
+
+                                return (
+                                    <div key={`thumbnail-${pageIndex}`}>
+                                        <div
+                                            className={classNames({
+                                                'rpv-thumbnail__item': true,
+                                                'rpv-thumbnail__item--selected': currentPage === pageIndex,
+                                            })}
+                                            role="button"
+                                            tabIndex={currentPage === pageIndex ? 0 : -1}
+                                            onClick={() => onJumpToPage(pageIndex)}
+                                        >
+                                            <ThumbnailContainer
+                                                doc={doc}
+                                                isActive={currentPage === pageIndex}
+                                                pageHeight={pageHeight}
+                                                pageIndex={pageIndex}
+                                                pageWidth={pageWidth}
+                                                rotation={rotation}
+                                                onActive={scrollToThumbnail}
+                                            />
+                                        </div>
+                                        <div className="rpv-thumbnail__index">{label}</div>
                                     </div>
-                                    <div className="rpv-thumbnail__index">
-                                        {index < numLabels && labels[index] ? labels[index] : index + 1}
-                                    </div>
-                                </div>
-                            ))}
+                                );
+                            })}
                     </div>
                 );
             }}