+++ packages/thumbnail/src/ThumbnailListWithStore.tsx
@@ -11,11 +11,13 @@ import { Spinner, useIsomorphicLayoutEffect } from '@react-pdf-viewer/core';
 import type { PdfJs, Store, StoreHandler } from '@react-pdf-viewer/core';
 
 import { ThumbnailList } from './ThumbnailList';
+import type { RenderCurrentPageLabel } from './types/RenderCurrentPageLabelProps';
 import type { StoreProps } from './types/StoreProps';
 
 export const ThumbnailListWithStore: React.FC<{
+    renderCurrentPageLabel?: RenderCurrentPageLabel;
     store: Store<StoreProps>;
-}> = ({ store }) => {
+}> = ({ renderCurrentPageLabel, store }) => {
     const [currentDoc, setCurrentDoc] = React.useState(store.get('doc'));
     const [currentPage, setCurrentPage] = React.useState(store.get('currentPage') || 0);
     const [pageHeight, setPageHeight] = React.useState(store.get('pageHeight') || 0);
@@ -77,6 +79,7 @@ export const ThumbnailListWithStore: React.FC<{
             doc={currentDoc}
             pageHeight={pageHeight}
             pageWidth={pageWidth}
+            renderCurrentPageLabel={renderCurrentPageLabel}
             rotation={rotation}
             onJumpToPage={jump}
         />