+++ packages/toolbar/src/Toolbar.tsx
@@ -8,7 +8,7 @@
 
 import * as React from 'react';
 
-import { defaultToolbar } from './defaultToolbar';
+import { DefaultToobar } from './DefaultToobar';
 import type { ToolbarSlot } from './types/ToolbarSlot';
 
 export type RenderToolbarSlot = (toolbarSlot: ToolbarSlot) => React.ReactElement;
@@ -21,6 +21,6 @@ export const Toolbar: React.FC<{
     children?: RenderToolbarSlot;
     slot: ToolbarSlot;
 }> = ({ children, slot }) => {
-    const render = children || defaultToolbar;
+    const render = children || DefaultToobar;
     return render(slot);
 };