+++ packages/toolbar/src/index.d.ts
@@ -110,8 +110,13 @@ export interface ToolbarSlot {
     ZoomOutMenuItem(props: ZoomMenuItemProps): React.ReactElement;
 }
 
+export type TransformToolbarSlot = (toolbarSlot: ToolbarSlot) => ToolbarSlot;
+
 // Plugin
 export interface ToolbarPlugin extends Plugin {
+    renderDefaultToolbar: (
+        transformToolbarSlot: TransformToolbarSlot
+    ) => (defaultToolbarSlot: ToolbarSlot) => React.ReactElement;
     Toolbar: (props: ToolbarProps) => React.ReactElement;
     // Plugins instance
     readonly fullScreenPluginInstance: FullScreenPlugin;